0.7.1开发版本bug的修复

+++ source/platonWallet/Modules/TransferService/TransactionService+Centralization.swift
@@ -16,7 +16,7 @@ extension TransactionService {
     
     public func getBatchTransaction(
         addresses: [String],
-        beginSequence: Int,
+        beginSequence: Int64,
         listSize: Int,
         direction: String,
         completion: PlatonCommonCompletion?) {
